Refactor: relocate adapter_automation module and fix imports
