+++ adapter_automation/ADAPTER_GENERATION.md
@@ -1,10 +1,10 @@
-# BioCypher Robust Adapter Wizard
+# BioCypher Adapter  Automation
 
 An automated pipeline for generating BioCypher Knowledge Graph adapters using LLMs (OpenRouter). This tool streamlines the process of transforming raw biological data files into production-ready Python adapters with minimal manual coding.
 
 ## Overview
 
-The Robust Adapter Wizard automates the following workflow:
+The Adapter automates the following workflow:
 1.  **Data Inspection**: Detects delimiters, headers, and samples data structure from local files.
 2.  **Semantic Mapping**: Uses LLM to infer the meaning of data columns and propose relationship properties.
 3.  **Specification Generation**: Creates a "Logic Blueprint" (YAML) that defines join logic, ID normalization, and implementation steps.
@@ -29,10 +29,10 @@ The Robust Adapter Wizard automates the following workflow:
 Run the interactive wizard from the project root:
 
 ```bash
-uv run python3 schema_generator/interactive_adapter_cli.py
+uv run python3 adapter_automation/interactive_adapter_cli.py
 ```
 
-### Wizard Steps
+### Steps
 
 1.  **Select Configuration**: 
     - Choose an existing adapter from `hsa_adapters_config_sample.yaml`.
@@ -50,12 +50,12 @@ uv run python3 schema_generator/interactive_adapter_cli.py
 
 - **`interactive_adapter_cli.py`**: The main user interface.
 - **`llm_adapter_generator.py`**: Orchestrates the code generation process.
-- **`logic_inference.py`**: Handles complex joining logic for auxiliary files and biological identifiers.
+- **`logic_inference.py`**: Handles  joining logic for auxiliary files and biological identifiers.
 - **`code_fixer.py`**: A defensive layer that automatically detects and repairs syntax errors or hallucinations in the generated code.
 - **`source_inspector.py`**: Deterministic analysis of file structure and data types.
 
 ## Debugging
 
 If generation fails, the tool saves diagnostic data to the `debug_traces/` directory. These files contain the full prompts and raw LLM responses used during the failed attempt, allowing for detailed troubleshooting of logic or context errors.
-- if u got 0 output, make sure to check the generated adapter is used the correct processor method if any.
+- if u got 0 output, make sure to check the generated adapter is used the correct processor method and column filter  if any.
