+++ adapter_automation/llm_column_mapper.py
