+++ adapter_automation/source_inspector.py
